+++ admin/includes/view_all_users.php
@@ -30,38 +30,59 @@
             <?php
 
             if (isset($_GET['change_to_admin'])) {
-                $user_id = $_GET['change_to_admin'];
 
-                $query = "UPDATE users SET user_role = 'admin' WHERE user_id = {$user_id} ";
-                $admin_query = mysqli_query($connection, $query);
+                if (isset($_SESSION['user_role'])) {
+                    if ($_SESSION['user_role'] == 'admin') {
 
-                confirmQuery($admin_query);
+                        $user_id = mysqli_real_escape_string($connection, $_GET['change_to_admin']);
 
-                header("Location: users.php");
+                        $query = "UPDATE users SET user_role = 'admin' WHERE user_id = {$user_id} ";
+                        $admin_query = mysqli_query($connection, $query);
+
+                        confirmQuery($admin_query);
+
+                        header("Location: users.php");
+                    }
+                }
+
+
             }
 
 
             if (isset($_GET['change_to_sub'])) {
-                $user_id = $_GET['change_to_sub'];
 
-                $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = {$user_id} ";
-                $sub_query = mysqli_query($connection, $query);
+                if (isset($_SESSION['user_role'])) {
+                    if ($_SESSION['user_role'] == 'admin') {
 
-                confirmQuery($sub_query);
+                        $user_id = mysqli_real_escape_string($connection, $_GET['change_to_sub']);
 
-                header("Location: users.php");
+                        $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = {$user_id} ";
+                        $sub_query = mysqli_query($connection, $query);
+
+                        confirmQuery($sub_query);
+
+                        header("Location: users.php");
+                    }
+                }
+
             }
 
 
             if (isset($_GET['delete'])) {
-                $user_id = $_GET['delete'];
 
-                $query = "DELETE FROM users WHERE user_id = {$user_id} ";
-                $delete_query = mysqli_query($connection, $query);
+                if (isset($_SESSION['user_role'])) {
+                    if ($_SESSION['user_role'] == 'admin') {
+                        $user_id = mysqli_real_escape_string($connection, $_GET['delete']);
 
-                confirmQuery($delete_query);
+                        $query = "DELETE FROM users WHERE user_id = {$user_id} ";
+                        $delete_query = mysqli_query($connection, $query);
 
-                header("Location: users.php");
+                        confirmQuery($delete_query);
+
+                        header("Location: users.php");
+                    }
+                }
+
             }
